added most functionallity for the main API

+++ server/controllers/goalController.js
@@ -1,13 +1,15 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
+
 
 /**
  * @description Get goals
  * @route GET /api/goals/
  * @access Private
  */
-const getGoals = asyncHandler(async (_, res) => {
-  const goals = await Goal.find();
+const getGoals = asyncHandler(async (req, res) => {
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json({ goals });
 });
 
@@ -21,7 +23,10 @@ const setGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Please add a text field");
   }
-  const goal = await Goal.create({ text: req.body.text });
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  });
 
   res.status(200).json({ goal });
 });
@@ -39,6 +44,18 @@ const editGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -52,13 +69,26 @@ const editGoal = asyncHandler(async (req, res) => {
  * @access Private
  */
 const deleteGoal = asyncHandler(async (req, res) => {
-  const goal = await Goal.findById(req.params.id)
+  const goal = await Goal.findById(req.params.id);
   if (!goal) {
     res.status(400);
-    throw new Error("Goal now found")
+    throw new Error("Goal now found");
   }
-  await goal.deleteOne()
-  res.status(200).json({id: req.params.id});
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.id.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await goal.deleteOne();
+  res.status(200).json({ id: req.params.id });
 });
 
 module.exports = { getGoals, setGoal, editGoal, deleteGoal };
